added 8443 -> 443 port redirect to ferm.conf
This was an oversight, but now the frontend can terminate https
connections whichs helps the go tool which first expects an https
connection, and on most platforms will wait quite a while for the
connection timeout before falling back to http, even though there is no
one listening on the port. That seems like the wrong behavior, but this
works around it anyways.